requirements.txt
Flask>=2.0.0
pandas>=1.3.0
PyPDF2>=3.0.0
pdfminer.six>=20201018
python-dotenv>=0.15.0
openai>=0.27.0